requirements.txt
gradio
selenium
beautifulsoup4
Pillow
google-generativeai
python-dotenv
asyncio
aiohttp
urllib3